+++ src/index.ts
@@ -1,0 +1,2 @@
+export type { Renderable } from './Renderable';
+export * as Renderer2D from './Renderer2D';