+++ vite.config.ts
@@ -1,0 +1,23 @@
+import { defineConfig } from 'vite';
+import dts from 'vite-plugin-dts';
+
+export default defineConfig({
+  plugins: [
+    dts({
+      insertTypesEntry: true,
+    }),
+  ],
+  build: {
+    lib: {
+      entry: 'src/index.ts',
+      name: 'math-ts',
+      fileName: 'index',
+      formats: ['es'],
+    },
+    /*
+    rollupOptions: {
+      external: [],
+    },
+    */
+  },
+});